feat(migration): reafactor migrations and add new migration steps
